Remove old workflows, add new one for core releases
