+++ .github/actions/checks/create-pull-request-check/action.yml
@@ -16,6 +16,9 @@ inputs:
   DETAILS_URL: 
     description: "Details url"
     required: true
+  REPOSITORY: 
+    description: "On witch repository add check "
+    required: false
 
 outputs:
   CHECK_ID:
@@ -30,11 +33,12 @@ runs:
       env:
         GH_TOKEN: ${{ github.token }}
         PR_HEAD: ${{ inputs.PR_HEAD }}
+        REPOSITORY: ${{ inputs.REPOSITORY  || "saleor-dashboard" }}
       run: |
         gh api \
         -H "Accept: application/vnd.github+json" \
         -H "X-GitHub-Api-Version: 2022-11-28" \
-        "/repos/saleor/saleor-dashboard/pulls?head=saleor:${PR_HEAD}" > data.json
+        "/repos/saleor/${REPOSITORY}/pulls?head=saleor:${PR_HEAD}" > data.json
         head_sha_with_quotes=$(jq '.[0] | .head.sha' data.json)
         echo "PULL_REQUEST_HEAD_SHA=${head_sha_with_quotes//\"/}"  >> "$GITHUB_ENV"
 
@@ -47,6 +51,8 @@ runs:
         NAME: ${{ inputs.CHECK_NAME }}
         STATUS: ${{ inputs.STATUS }}
         TITLE: ${{ inputs.TITLE }}
+        REPOSITORY: ${{ inputs.REPOSITORY  || "saleor-dashboard" }}
+
       run: |
         payload=$(\
         jq --null-input \
@@ -73,5 +79,5 @@ runs:
         -H "Accept: application/vnd.github+json" \
         -H "Authorization: Bearer $GH_TOKEN" \
         -H "X-GitHub-Api-Version: 2022-11-28" \
-        https://api.github.com/repos/saleor/saleor-dashboard/check-runs \
+        https://api.github.com/repos/$REPOSITORY/saleor-dashboard/check-runs \
         -d "$payload" | jq -r '.id') ; echo "check_id=$check_id" >> "$GITHUB_OUTPUT"
