+++ .github/actions/checks/update-pull-request-check/action.yml
@@ -22,6 +22,9 @@ inputs:
   DETAILS_URL:
     description: "Url with details to check"
     required: true
+  REPOSITORY: 
+    description: "Repository"
+    required: false
 
 runs:
   using: composite
@@ -37,6 +40,7 @@ runs:
         TITLE: ${{ inputs.TITLE }}
         SUMMARY: ${{ inputs.SUMMARY }}
         DETAILS_URL: ${{ inputs.DETAILS_URL }}
+        REPOSITORY: ${{ inputs.REPOSITORY  || "saleor-dashboard" }}
       run: |
       
         payload=$(\
@@ -65,5 +69,5 @@ runs:
         -H "Accept: application/vnd.github+json" \
         -H "Authorization: Bearer $GH_TOKEN" \
         -H "X-GitHub-Api-Version: 2022-11-28" \
-        "https://api.github.com/repos/saleor/saleor-dashboard/check-runs/$CHECK_ID" \
+        "https://api.github.com/repos/$REPOSITORY/saleor-dashboard/check-runs/$CHECK_ID" \
         -d "$payload"
